requirements_final_gpu.txt
# AI Answer Evaluation System - CUDA 12.1 GPU Requirements
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
pytesseract==0.3.10
fpdf2==2.7.9
reportlab==4.0.7
numpy==1.26.2
scikit-learn==1.3.2
chromadb==0.4.18
sentence-transformers==2.7.0
huggingface-hub==0.24.0
requests==2.31.0
accelerate==0.34.0
transformers==4.45.0
addict==2.4.0
timm==0.9.12
einops==0.7.0
ollama==0.1.6

--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
